:recycle: Added modular GetCardDetails.kt and TextProcessor.kt
