+++ app/src/main/java/in/surajsau/jisho/domain/cardreader/processor/TextProcessor.kt
@@ -1,0 +1,7 @@
+package `in`.surajsau.jisho.domain.cardreader.processor
+
+import com.google.mlkit.vision.text.Text
+
+interface TextProcessor<T> {
+    fun process(textResult: Text): T
+}
